Update CollectiblesList component
