+++ src/components/CollectiblesList/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "CollectiblesList",
+  "version": "0.0.0",
+  "main": "./CollectiblesList.js"
+}
